DSP_project/requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
Pillow
scikit-learn
scipy
deap
pytest
pickle
